refactor: use same layour for desktop and mobile
