+++ frontend/components.d.ts
@@ -23,7 +23,7 @@ declare module 'vue' {
     ChangePasswordModal: typeof import('./src/components/Modals/ChangePasswordModal.vue')['default']
     CopyControl: typeof import('./src/components/Controls/CopyControl.vue')['default']
     DashboardLayout: typeof import('./src/components/DashboardLayout.vue')['default']
-    DesktopLayout: typeof import('./src/components/DesktopLayout.vue')['default']
+    DefaultLayout: typeof import('./src/components/DefaultLayout.vue')['default']
     EditAliasModal: typeof import('./src/components/Modals/EditAliasModal.vue')['default']
     EditPhotoModal: typeof import('./src/components/Modals/EditPhotoModal.vue')['default']
     EmojiPicker: typeof import('./src/components/EmojiPicker.vue')['default']
@@ -40,7 +40,6 @@ declare module 'vue' {
     MailLogo: typeof import('./src/components/Icons/MailLogo.vue')['default']
     MailThread: typeof import('./src/components/MailThread.vue')['default']
     MailThreadPlaceholder: typeof import('./src/components/MailThreadPlaceholder.vue')['default']
-    MobileLayout: typeof import('./src/components/MobileLayout.vue')['default']
     MultiselectInputControl: typeof import('./src/components/Controls/MultiselectInputControl.vue')['default']
     NoMails: typeof import('./src/components/Icons/NoMails.vue')['default']
     ProfileSettings: typeof import('./src/components/Settings/ProfileSettings.vue')['default']
